- check if directory exists before deleting

+++ RocksmithToolkitLib/PsarcLoader/PsarcPackager.cs
@@ -63,7 +63,7 @@ namespace RocksmithToolkitLib.PsarcLoader
         protected virtual void Dispose(Boolean disposing)
         {
             if (disposing)
-                if (_deleteOnClose)
+                if (_deleteOnClose && Directory.Exists(packageDir))
                     IOExtension.DeleteDirectory(packageDir);
         }
